Made some scripts for testing some sensors
higro.ino is for the higrometer, which is a simple
analog read on the input.

sonar.ino is for the sonars, which needs a library called
Ultrasonic because it is based on the delay between the
trigger and echo signals.

sensors.ino now have the functionalities of the higrometer
and the sonars.